Adding explanation and alt texts
It's a pity that the images used here are annotated and will therefore
not be accessible. At the very least I added alt texts so that some
meaning can be recovered.

Since this page assumes a novice I've went ahead and added some more
explanation to different sections. I'm happy to see them upstreamed to
Codeberg.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>