docs: developer: workflow: requirement for Forgejo database migrations
While feature branches may have different requirements for the Forgejo
tables, they are constrained by the migration logic that requires they
happen in sequence and are listed in the same list. If each feature
branch changes the `forgejo_migrations/migrate.go` file independently,
it is bound to conflict with other feature branches trying to do the
same.

To avoid this situation it is proposed that once a migration
developped and tested in a freature branch is final, it is isolated in
a commit and moved to the `forgejo-development` branch. This can
happen during a rebase or with a pull request.

It is probably better that all migrations are ultimately squashed in a
single commit in the `forgejo-development` branch instead of an ever
growing list of commits. But this is up to the person in charge of the
migrations and since the directory is owned by forgejo, there is no
risk of conflict.