admins: actions: one token can register multiple runners
See https://codeberg.org/forgejo/forgejo/commit/9b698362a333de2c388499f1a64d39545b0263bd

Also remove the token line from the .runner file. It is confusing and
difficult to explain because it is different from the registration
token. It really is a shared secret between the runner and the Forgejo
instance that has a purpose which is entirely different.