{admin,user}: actions: runner 3.1+ support lxc + self-hosted
The upcoming 3.1 release of the Forgeo runner will allow to specify
which template and release must be must be used for a LXC
container. It also defines the lxc:// scheme as distinct from the
host:// scheme.

The host:// scheme is documented to be used for running jobs directly
from the host, which was not possible with the Forgejo runner versions
prior to 3.1