developer: infrastructure: hetzner{02,03} root file system backups
No information in the root filesystem is really worth saving as it
can be recreated from this documentation or even from scratch manually
if needs be. But it is cheap to keep that backup and it is worth the
effort if it can save a few hours of work when an accident happen.

The more sophisticated way to do it would be to setup DRBD with the
root file system. But this is not straightforward and to ensure it
won't get the admin in trouble when an accident happen is tricky. It
is the one time when every bit of complication can quickly become a
blocker or an unecessary burden.