admin/customization: No need to be a developer for custom branding (#427)
Preview: https://forgejo.codeberg.page/@docs_pull_427/docs/next/admin/customization/

---

The previous version of this document insinuated that you need to be
familiar with Forgejo's source code and compile Forgejo from source
for custom branding changes. As of the most recent Forgejo version,
this is not necessarily true, as custom files can be served from
the CustomPath.

Administrators that were not developers therefore dismissed the
other guide and its contents without reading this, and I was one
of them. I thought it would make sense for this file to provide
information on where to put a custom logo. This change does not
fix it (this will be done later), but it will reduce the confusion
for now.

This warning confused me personally and other Forgejo users, as I
had one person tell me IRL that they thought that it was not possible
to serve a custom logo.

Reviewed-on: https://codeberg.org/forgejo/docs/pulls/427
Reviewed-by: Earl Warren <earl-warren@noreply.codeberg.org>
Co-authored-by: Panagiotis "Ivory" Vasilopoulos <git@n0toose.net>
Co-committed-by: Panagiotis "Ivory" Vasilopoulos <git@n0toose.net>