admin: Reword, format consistency
This changes terminal command blocks to always use the multiline formatting,
as the leading octothorpe looks very strange when displaying a command inline.
Two consecutive inline commands at the start was very difficult to read in the
generated html. See:
https://forgejo.org/docs/latest/admin/installation-binary/

`HTTP_PORT` has been highlighted as a code term. This also means we don't need
to worry about another underscore in the same line being added and interpreted
as syntax.

Simplifies and fixes sentence structure in a few places. I've seen arguments
for and against the use of "just" in technical writing, but I think here it's
better without?

Reworded some paragraphs so that the command comes at the end.