Clarify API token usage (#939)
The config section can be easily missed when someone intends to only use the token as an env var. But that section contains the crucial bit of information about formatting the token.
This splits the credentials info from configuration description and makes the format explicit. While at it, the necessary permission is highlighted.

This should fix Codeberg/Community#1701

Co-authored-by: Earl Warren <earl-warren@noreply.codeberg.org>
Reviewed-on: https://codeberg.org/forgejo/docs/pulls/939
Reviewed-by: Earl Warren <earl-warren@noreply.codeberg.org>
Co-authored-by: dcz_pf <cobepfac.dcz@porcupinefactory.org>
Co-committed-by: dcz_pf <cobepfac.dcz@porcupinefactory.org>