feat: add profiling documentation
Add documentation for how you can and do diagnostics/profiling in
Forgejo, it covers the methods that Forgejo offer for this,
nuances of the methods and potential footguns. I don't expect the
average forgejo adminstrator to be able to profile their Forgejo
instance (hence the warning in the first paragraph about contacting the
Forgejo developers) but for those brave enough this documentation should
be a solid foundation to get started.

This is complementatory the Forgejo PR about improving the diagnostics: