Fix broken links in runner-installation.md
I've fixed broken links, and where I was sure was appropriate, updated
the version in the link.

I did not change the version in the links in the `OCI image
installation` section, as I was unsure if I might actually break
something by changing the commands or the docker-compose.yaml. I haven't
had the opportunity to try and test that method out yet.