fix: GITHUB_TOKEN is not needed for PR previews (#1172)
GITHUB_TOKEN is not needed when building a preview but will be set
with write permissions to the repository because it runs from a
pull_request_target. Because it explicitly checkout the head of the
PR, scripts found in the repository are run from the head of the pull
request and not from the base branch.

A malicious pull request could modify the script and push to the
repository, reading that read-write token from the environment.

Overriding the environment variable on each run step prevents that
from happening.

Reviewed-on: https://codeberg.org/forgejo/docs/pulls/1172
Co-authored-by: Earl Warren <contact@earl-warren.org>
Co-committed-by: Earl Warren <contact@earl-warren.org>