Clarify confusion about pull request template files (#1231)
It was not mentioned in the docs where to put a pull request template file in order for it to work.
I saw a working example in `forgejo`'s [own repo](https://code.forgejo.org/forgejo/pr-and-issue-templates/src/branch/main/.forgejo/pull_request_template.md) and propose that this should be clarified on the [current documentation site](https://forgejo.org/docs/latest/user/issue-pull-request-templates/), which is heavily focused on issue template files.

Reviewed-on: https://codeberg.org/forgejo/docs/pulls/1231
Reviewed-by: Lucas <sclu1034@noreply.codeberg.org>
Reviewed-by: Robert Wolff <mahlzahn@posteo.de>
Co-authored-by: Isabel Lesjak <isabel.lesjak@geizhals.at>
Co-committed-by: Isabel Lesjak <isabel.lesjak@geizhals.at>