Reorganize and partially rewrite the actions documentation (#1094)
The Forgejo Actions documentation is not ideal. Currently, most of it is [one huge page](https://forgejo.org/docs/latest/user/actions/) with unclear organization and little separation between beginner concepts and advanced details. This PR aims to reorganize the currently written text of the actions documentation into a better organized structure, and to supplement it with new/rewritten text where necessary.

Reviewed-on: https://codeberg.org/forgejo/docs/pulls/1094
Reviewed-by: Otto <otto@codeberg.org>
Reviewed-by: Earl Warren <earl-warren@noreply.codeberg.org>
Co-authored-by: Vivian <me@vivstuff.gay>
Co-committed-by: Vivian <me@vivstuff.gay>