docs: document git credentials behavior with checkout action (#1375)
After discovering an issue with git credentials (see https://codeberg.org/forgejo/forgejo/issues/8839), I wanted to document this somewhere in case people are running into the same issue.
For me, that looks like a topic that could happen sometimes.

Not sure if this is the best place though, if not, feel free to send me elsewhere.

Reviewed-on: https://codeberg.org/forgejo/docs/pulls/1375
Reviewed-by: Earl Warren <earl-warren@noreply.codeberg.org>
Co-authored-by: Vanessa Gaube <dev@vanessagaube.de>
Co-committed-by: Vanessa Gaube <dev@vanessagaube.de>