docs: describe common configurations for Actions to interact w/ docker (#1421)
One of the most common struggles I've seen on social media with using Forgejo Actions is that people are unable to do `docker build` within an action.  The goal of this PR is to document the solutions for self-hosted runners that can allow this configuration, and provide all the information necessary for a self-hosted administrator to choose the right configuration for them.

Social media references, for example:
- https://old.reddit.com/r/forgejo/comments/1n27luv/trouble_building_docker_images_with_forgejo/
- https://notes.jamesravey.me/Software/Forgejo#docker-in-docker
- https://bsky.app/profile/rachel.transitory.social.ap.brid.gy/post/3lipakzkh4vn2
- https://dbushell.com/2025/08/15/self-hosted-forgejo-actions-runner/
    - "Update: after researching this setup more, I realise mounting the host Docker socket is a security risk. A malicious 3rd-party action could gain access to host files."

Reviewed-on: https://codeberg.org/forgejo/docs/pulls/1421
Reviewed-by: Earl Warren <earl-warren@noreply.codeberg.org>
Reviewed-by: Lucas <sclu1034@noreply.codeberg.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>