docs: detailed documentation on options for securing a Forgejo Runner installation (#1429)
Adds a new "Securing Forgejo Actions Deployments" document to the "Forgejo Actions administrator guide".  This document attempts to paint a picture of the software configuration and system configuration options that will contribute to managing the risk in a Forgejo Actions deployment.

This PR builds upon the documentation in #1421; only the "docs: detailed documentation on securing a Forgejo Runner installation" commit is an addition relative to that PR at this time.

Reviewed-on: https://codeberg.org/forgejo/docs/pulls/1429
Reviewed-by: Earl Warren <earl-warren@noreply.codeberg.org>
Reviewed-by: Lucas <sclu1034@noreply.codeberg.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>