fix: incorrect that 'inputs' context is available within 'inputs' (#1628)
When evaluating `on.workflow_call.inputs.<input-id>.default` in a Forgejo Action, it is not possible to reference `inputs` -- the `inputs` for the current workflow are being *defined* during the evaluation of this context, and this would be a circular reference.  The workflow schema check file with this mistake should be fixed in https://code.forgejo.org/forgejo/runner/pulls/1211.

Reviewed-on: https://codeberg.org/forgejo/docs/pulls/1628
Reviewed-by: Gusted <gusted@noreply.codeberg.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>